+++ dbentry/base/admin.py
@@ -125,10 +125,8 @@ class MIZModelAdmin(AutocompleteMixin, MIZAdminSearchFormMixin, admin.ModelAdmin
                         get_fields_and_lookups(self.model, field)
                 except (exceptions.FieldDoesNotExist, exceptions.FieldError) as e:
                     errors.append(
-                        # TODO: use f-strings
                         checks.Error(
-                            "fieldset '%s' contains invalid item: '%s'. %s" % (
-                                fieldset_name, field, e.args[0]),
+                            f"fieldset '{fieldset_name}' contains invalid item: '{field}'. {e.args[0]}",
                             obj=self.__class__
                         )
                     )
@@ -138,11 +136,12 @@ class MIZModelAdmin(AutocompleteMixin, MIZAdminSearchFormMixin, admin.ModelAdmin
         return request.user.is_superuser
 
     def get_queryset(self, request: HttpRequest) -> QuerySet:
-        qs = super().get_queryset(request)
-        annotations = self.get_changelist_annotations()
-        if annotations:
-            return qs.annotate(**annotations)
-        return qs
+        queryset = super().get_queryset(request)
+        # overview() adds annotations and queryset optimizations for the
+        # changelist view:
+        if not hasattr(queryset, "overview"):
+            return queryset
+        return queryset.overview()
 
     def get_changelist(self, request: HttpRequest, **kwargs: Any) -> Type[MIZChangeList]:
         return MIZChangeList
@@ -425,10 +424,6 @@ class MIZModelAdmin(AutocompleteMixin, MIZAdminSearchFormMixin, admin.ModelAdmin
         if isinstance(form.instance, ComputedNameModel):
             # Update the instance's _name now. save_model was called earlier.
             form.instance.update_name(force_update=True)
-
-    def get_changelist_annotations(self) -> dict:
-        """Return annotations necessary for the changelist queryset."""
-        return {}
 
     def get_search_results(
             self,