+++ dbentry/base/models.py
@@ -22,13 +22,19 @@ class BaseModel(models.Model):
           output of __str__().
         - ``create_field`` (str): the name of the field for the dal
           autocomplete object creation.
-        - ``exclude_from_str`` (list): list of field names to be excluded from
+        - ``exclude_from_str`` (tuple): tuple of field names to be excluded from
           the default __str__() implementation.
+        - ``select_related`` (tuple): tuple of ForeignKey field names. Used in
+          optimizations for the overview queryset.
+        - ``prefetch_related`` (tuple): tuple of M2M field names. Used in
+          optimizations for the overview queryset.
     """
 
     name_field: str = ''
     create_field: str = ''
-    exclude_from_str: list = ['beschreibung', 'bemerkungen', '_fts']
+    exclude_from_str: tuple = ('beschreibung', 'bemerkungen', '_fts')
+    select_related: tuple = ()
+    prefetch_related: tuple = ()
 
     objects = MIZQuerySet.as_manager()
 
@@ -46,18 +52,11 @@ class BaseModel(models.Model):
         if self.name_field:
             result = str(opts.get_field(self.name_field).value_from_object(self))
         else:
-            model_fields = get_model_fields(
-                opts.model,
-                foreign=False,
-                m2m=False,
-                exclude=self.exclude_from_str
-            )
-            # TODO: replace the above with the below to remove the get_model_fields call:
-            # model_fields = [
-            #     f for f in opts.get_fields()
-            #     if f.concrete
-            #     and not (f.primary_key or f.is_relation or f.name in self.exclude_from_str)
-            # ]
+            model_fields = [
+                f for f in opts.get_fields()
+                if f.concrete
+                and not (f.primary_key or f.is_relation or f.name in self.exclude_from_str)
+            ]
             result = " ".join(
                 [
                     str(fld.value_from_object(self))
@@ -81,6 +80,31 @@ class BaseModel(models.Model):
             )
         # noinspection PyUnresolvedReferences
         return self._meta.model.objects.filter(pk=self.pk)
+
+    @staticmethod
+    def get_overview_annotations() -> dict:
+        """Return a dictionary of queryset annotations required for a changelist view."""
+        return {}
+
+    @classmethod
+    def overview(cls, queryset, *annotations):
+        """
+        Add annotations and optimizations useful for an overview over objects
+        of this model to the given queryset.
+
+        If `annotations` is given, only apply matching annotations.
+        """
+        if select_related := cls.select_related:
+            queryset = queryset.select_related(*select_related)
+        if prefetch_related := cls.prefetch_related:
+            queryset = queryset.prefetch_related(*prefetch_related)
+        if all_annotations := cls.get_overview_annotations():
+            if annotations:
+                _annotations = {k: v for k, v in all_annotations.items() if k in annotations}
+            else:
+                _annotations = all_annotations
+            queryset = queryset.annotate(**_annotations)
+        return queryset
 
     class Meta:
         abstract = True